+++ examples/ios/swift-2.0/Simple/AppDelegate.swift
@@ -57,7 +57,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Save your object
         realm.beginWrite()
         realm.add(mydog)
-        realm.commitWrite()
+        try! realm.commitWrite()
 
         // Query
         let results = realm.objects(Dog).filter(NSPredicate(format:"name contains 'x'"))
@@ -72,7 +72,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         person.name = "Tim"
         person.dogs.append(mydog)
 
-        realm.write {
+        try! realm.write {
             realm.add(person)
         }